moving transform functions back to filter
