+++ metron-interface/metron-alerts/src/app/utils/utils.ts
@@ -204,41 +204,4 @@ export class Utils {
 
     return {toDate: toDate, fromDate: fromDate};
   }
-
-  public static transformModelToControlValue(model: PcapRequest): PcapFilterFormValue {
-    const startTimeStr = moment(model.startTimeMs > 0 ? model.startTimeMs : DEFAULT_START_TIME).format(DEFAULT_TIMESTAMP_FORMAT);
-    let endTimeStr = moment(model.endTimeMs).format(DEFAULT_TIMESTAMP_FORMAT);
-    if (isNaN((new Date(model.endTimeMs).getTime()))) {
-      endTimeStr = moment(DEFAULT_END_TIME).format(DEFAULT_TIMESTAMP_FORMAT);
-    } else {
-      endTimeStr = moment(model.endTimeMs).format(DEFAULT_TIMESTAMP_FORMAT);
-    }
-
-    return {
-      startTime: startTimeStr,
-      endTime: endTimeStr,
-      ipSrcAddr: model.ipSrcAddr,
-      ipDstAddr: model.ipDstAddr,
-      ipSrcPort: model.ipSrcPort ? String(model.ipSrcPort) : '',
-      ipDstPort: model.ipDstPort ? String(model.ipDstPort) : '',
-      protocol: model.protocol,
-      includeReverse: model.includeReverse,
-      packetFilter: model.packetFilter
-    };
-  }
-
-  public static transformControlValueToModel(control: FormGroup): PcapRequest {
-    const pcapRequest = new PcapRequest();
-    pcapRequest.startTimeMs = new Date(control.value.startTime).getTime();
-    pcapRequest.endTimeMs = new Date(control.value.endTime).getTime();
-    pcapRequest.ipSrcAddr = control.value.ipSrcAddr;
-    pcapRequest.ipDstAddr = control.value.ipDstAddr;
-    pcapRequest.ipSrcPort = control.value.ipSrcPort;
-    pcapRequest.ipDstPort = control.value.ipDstPort;
-    pcapRequest.protocol =  control.value.protocol;
-    pcapRequest.includeReverse = control.value.includeReverse;
-    pcapRequest.packetFilter = control.value.packetFilter;
-    return pcapRequest;
-  }
-
 }
